requirements.txt
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
filelock==3.17.0
frozenlist==1.5.0
futures==3.0.5
googlesearch-python==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
msgpack==1.1.0
openai==1.61.1
packaging==24.2
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
ray==2.42.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
